+++ NEWS.md
@@ -1,6 +1,6 @@
 # PAnEn 4.3.0
 
-- Added `AnEnSSEMS` for matching stations
+- Added `AnEnSSEMS` for matching stations. This addition has also been ported to `RAnEn`.
 
 # PAnEn 4.2.7